requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
requests>=2.31.0
